requirements.txt
numpy
scikit-learn
sentence-transformers
faiss-cpu
pytest
termcolor
pyfiglet
chromadb
